requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
pyyaml==6.0.1
dvc==3.48.4
fastapi==0.111.0
uvicorn==0.30.1
